move lib tests to tests.rs
